Split Multiverse file in two.
