+++ sbnana/CAFAna/Core/EnsembleRatio.cxx
@@ -2,7 +2,7 @@
 
 #include "sbnana/CAFAna/Core/EnsembleSpectrum.h"
 
-#include "sbnana/CAFAna/Core/Multiverse.h"
+#include "sbnana/CAFAna/Core/FitMultiverse.h"
 #include "sbnana/CAFAna/Core/Utilities.h"
 
 #include "cafanacore/UtilsExt.h"
